Create README.md
Description: program that will ask the user to enter a line of text to be compressed. Iterate over the string from beginning to end. For each
character in the string, if the character is the same as the previous one then increment a counter,
otherwise print out the previous count and character before resetting the counter. When you run out
of characters in the string make sure you've output the final pair.